验证码服务端, 业务后台基本完工, 还差前端

+++ src/SimCaptcha/Models/PointPosModel.cs
@@ -6,8 +6,8 @@ namespace SimCaptcha.Models
 {
     public class PointPosModel
     {
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
     }
 }
